Second commit

+++ src/controller/AttaqueGestion.java
@@ -4,12 +4,12 @@ import model.Pokemon;
 
 public class AttaqueGestion {
 	
-	public static void GestionAttaque(Pokemon pPokemon1, Pokemon pPokemon2, float attaque) {
+	public static void GestionAttaque(Pokemon pPokemon1, Pokemon pPokemon2, double attaque) {
 		
 		if (pPokemon1 != pPokemon2) {
 			
-			float energieAvantAttaque = pPokemon1.getEnergie();
-			float energieApresAttaque = (attaque > energieAvantAttaque) ? (float)0 : (energieAvantAttaque - attaque);
+			double energieAvantAttaque = pPokemon1.getEnergie();
+			double energieApresAttaque = (attaque > energieAvantAttaque) ? 0.0 : (energieAvantAttaque - attaque);
 			pPokemon1.setEnergie(energieApresAttaque);
 			
 			System.out.println(pPokemon1.getNom() + " a " + energieAvantAttaque + " energie. " + pPokemon2.getNom() + " vient de l'attaquer en lui faisant " + attaque +" dégats.");
