+++ src/model/PokemonFeu.java
@@ -3,13 +3,13 @@ package model;
 public class PokemonFeu extends Pokemon {
 
 	private int nbAiles;
-	private float taille;
+	private double taille;
 	private int nbBouledeFeu;
-	private float puissanceFeu;
+	private double puissanceFeu;
 	
 	
-	public PokemonFeu(String pNom, float pEnergie, float pPoids, int nbAiles, float taille, int nbBouledeFeu,
-			float puissanceFeu) {
+	public PokemonFeu(String pNom, double pEnergie, double pPoids, int nbAiles, double taille, int nbBouledeFeu,
+			double puissanceFeu) {
 		super(pNom, pEnergie, pPoids);
 		this.nbAiles = nbAiles;
 		this.taille = taille;
@@ -22,7 +22,7 @@ public class PokemonFeu extends Pokemon {
 		return nbAiles;
 	}
 
-	public float getTaille() {
+	public double getTaille() {
 		return taille;
 	}
 
@@ -30,7 +30,7 @@ public class PokemonFeu extends Pokemon {
 		return nbBouledeFeu;
 	}
 
-	public float getPuissanceFeu() {
+	public double getPuissanceFeu() {
 		return puissanceFeu;
 	}
 
@@ -38,7 +38,7 @@ public class PokemonFeu extends Pokemon {
 		this.nbAiles = nbAiles;
 	}
 
-	public void setTaille(float taille) {
+	public void setTaille(double taille) {
 		this.taille = taille;
 	}
 
@@ -46,7 +46,7 @@ public class PokemonFeu extends Pokemon {
 		this.nbBouledeFeu = nbBouledeFeu;
 	}
 
-	public void setPuissanceFeu(float puissanceFeu) {
+	public void setPuissanceFeu(double puissanceFeu) {
 		this.puissanceFeu = puissanceFeu;
 	}
 
@@ -56,13 +56,13 @@ public class PokemonFeu extends Pokemon {
 				+ ", puissanceFeu=" + puissanceFeu + ", nom=" + nom + ", energie=" + energie + ", poids=" + poids + "]";
 	}
 	
-	public float VitesseVol() {
+	public double VitesseVol() {
 		return this.nbAiles * this.taille;
 	}
 	
 	public void attaque(Pokemon pPokemonAttaque) {
 		
-		float attaque = 15*this.VitesseVol();
+		double attaque = 15*this.VitesseVol();
 		super.attaque(pPokemonAttaque, this, attaque);
 		
 	}
